relocate stan/lang/rethrow_located.hpp
