+++ src/stan/model/model_header.hpp
@@ -8,7 +8,7 @@
 #include <stan/io/reader.hpp>
 #include <stan/io/writer.hpp>
 
-#include <stan/lang/rethrow_located.hpp>
+#include <stan/model/rethrow_located.hpp>
 #include <stan/model/model_base.hpp>
 #include <stan/model/model_base_crtp.hpp>
 #include <stan/model/prob_grad.hpp>
