+++ src/stan/model/standalone_functions_header.hpp
@@ -5,7 +5,7 @@
 
 #include <boost/random/additive_combine.hpp>
 #include <stan/io/program_reader.hpp>
-#include <stan/lang/rethrow_located.hpp>
+#include <stan/model/rethrow_located.hpp>
 #include <stan/model/indexing.hpp>
 
 #include <cmath>
