+++ src/test/unit/lang/generator/cholesky_factor_cov_var_test.cpp
@@ -227,7 +227,7 @@ TEST(lang, params_block_var_hpp_xform_inits) {
       "        current_statement_begin__ = 2;\n"
       "        if (!(context__.contains_r(\"cfcov_54\")))\n"
       "            "
-      "stan::lang::rethrow_located(std::runtime_error(std::string(\"Variable "
+      "stan::model::rethrow_located(std::runtime_error(std::string(\"Variable "
       "cfcov_54 missing\")), current_statement_begin__, prog_reader__());\n"
       "        vals_r__ = context__.vals_r(\"cfcov_54\");\n"
       "        pos__ = 0U;\n"
@@ -249,7 +249,7 @@ TEST(lang, params_block_var_hpp_xform_inits) {
       "            writer__.cholesky_factor_cov_unconstrain(cfcov_54);\n"
       "        } catch (const std::exception& e) {\n"
       "            "
-      "stan::lang::rethrow_located(std::runtime_error(std::string(\"Error "
+      "stan::model::rethrow_located(std::runtime_error(std::string(\"Error "
       "transforming variable cfcov_54: \") + e.what()), "
       "current_statement_begin__, prog_reader__());\n"
       "        }\n");
@@ -259,7 +259,7 @@ TEST(lang, params_block_var_hpp_xform_inits) {
       "        current_statement_begin__ = 3;\n"
       "        if (!(context__.contains_r(\"cfcov_33\")))\n"
       "            "
-      "stan::lang::rethrow_located(std::runtime_error(std::string(\"Variable "
+      "stan::model::rethrow_located(std::runtime_error(std::string(\"Variable "
       "cfcov_33 missing\")), current_statement_begin__, prog_reader__());\n"
       "        vals_r__ = context__.vals_r(\"cfcov_33\");\n"
       "        pos__ = 0U;\n"
@@ -281,7 +281,7 @@ TEST(lang, params_block_var_hpp_xform_inits) {
       "            writer__.cholesky_factor_cov_unconstrain(cfcov_33);\n"
       "        } catch (const std::exception& e) {\n"
       "            "
-      "stan::lang::rethrow_located(std::runtime_error(std::string(\"Error "
+      "stan::model::rethrow_located(std::runtime_error(std::string(\"Error "
       "transforming variable cfcov_33: \") + e.what()), "
       "current_statement_begin__, prog_reader__());\n"
       "        }\n");
@@ -503,7 +503,7 @@ TEST(lang, transformed_params_block_var_hpp_log_prob) {
       "                        msg__ << \"Undefined transformed parameter: "
       "cfcov_54\" << \"(\" << j_1__ << \", \" << j_2__ << \")\";\n"
       "                        "
-      "stan::lang::rethrow_located(std::runtime_error(std::string(\"Error "
+      "stan::model::rethrow_located(std::runtime_error(std::string(\"Error "
       "initializing variable cfcov_54: \") + msg__.str()), "
       "current_statement_begin__, prog_reader__());\n"
       "                    }\n"
@@ -525,7 +525,7 @@ TEST(lang, transformed_params_block_var_hpp_log_prob) {
       "                        msg__ << \"Undefined transformed parameter: "
       "cfcov_33\" << \"(\" << j_1__ << \", \" << j_2__ << \")\";\n"
       "                        "
-      "stan::lang::rethrow_located(std::runtime_error(std::string(\"Error "
+      "stan::model::rethrow_located(std::runtime_error(std::string(\"Error "
       "initializing variable cfcov_33: \") + msg__.str()), "
       "current_statement_begin__, prog_reader__());\n"
       "                    }\n"
