+++ src/test/unit/lang/generator/function_var_test.cpp
@@ -216,7 +216,7 @@ TEST(lang, local_function_args_hpp) {
       "            *pstream__ << std::endl;\n"
       "        }\n"
       "    } catch (const std::exception& e) {\n"
-      "        stan::lang::rethrow_located(e, current_statement_begin__, "
+      "        stan::model::rethrow_located(e, current_statement_begin__, "
       "prog_reader__());\n"
       "        // Next line prevents compiler griping about no return\n"
       "        throw std::runtime_error(\"*** IF YOU SEE THIS, PLEASE REPORT A "
@@ -288,7 +288,7 @@ TEST(lang, local_function_args_hpp) {
       "        current_statement_begin__ = 7;\n"
       "        return stan::math::promote_scalar<fun_return_scalar_t__>(a1);\n"
       "    } catch (const std::exception& e) {\n"
-      "        stan::lang::rethrow_located(e, current_statement_begin__, "
+      "        stan::model::rethrow_located(e, current_statement_begin__, "
       "prog_reader__());\n"
       "        // Next line prevents compiler griping about no return\n"
       "        throw std::runtime_error(\"*** IF YOU SEE THIS, PLEASE REPORT A "
@@ -360,7 +360,7 @@ TEST(lang, local_function_args_hpp) {
       "        current_statement_begin__ = 11;\n"
       "        return stan::math::promote_scalar<fun_return_scalar_t__>(a2);\n"
       "    } catch (const std::exception& e) {\n"
-      "        stan::lang::rethrow_located(e, current_statement_begin__, "
+      "        stan::model::rethrow_located(e, current_statement_begin__, "
       "prog_reader__());\n"
       "        // Next line prevents compiler griping about no return\n"
       "        throw std::runtime_error(\"*** IF YOU SEE THIS, PLEASE REPORT A "
@@ -432,7 +432,7 @@ TEST(lang, local_function_args_hpp) {
       "        current_statement_begin__ = 15;\n"
       "        return stan::math::promote_scalar<fun_return_scalar_t__>(a3);\n"
       "    } catch (const std::exception& e) {\n"
-      "        stan::lang::rethrow_located(e, current_statement_begin__, "
+      "        stan::model::rethrow_located(e, current_statement_begin__, "
       "prog_reader__());\n"
       "        // Next line prevents compiler griping about no return\n"
       "        throw std::runtime_error(\"*** IF YOU SEE THIS, PLEASE REPORT A "
@@ -506,7 +506,7 @@ TEST(lang, local_function_args_hpp) {
       "        current_statement_begin__ = 19;\n"
       "        return stan::math::promote_scalar<fun_return_scalar_t__>(a4);\n"
       "    } catch (const std::exception& e) {\n"
-      "        stan::lang::rethrow_located(e, current_statement_begin__, "
+      "        stan::model::rethrow_located(e, current_statement_begin__, "
       "prog_reader__());\n"
       "        // Next line prevents compiler griping about no return\n"
       "        throw std::runtime_error(\"*** IF YOU SEE THIS, PLEASE REPORT A "
@@ -584,7 +584,7 @@ TEST(lang, local_function_args_hpp) {
       "        current_statement_begin__ = 23;\n"
       "        return stan::math::promote_scalar<fun_return_scalar_t__>(a5);\n"
       "    } catch (const std::exception& e) {\n"
-      "        stan::lang::rethrow_located(e, current_statement_begin__, "
+      "        stan::model::rethrow_located(e, current_statement_begin__, "
       "prog_reader__());\n"
       "        // Next line prevents compiler griping about no return\n"
       "        throw std::runtime_error(\"*** IF YOU SEE THIS, PLEASE REPORT A "
@@ -662,7 +662,7 @@ TEST(lang, local_function_args_hpp) {
       "        current_statement_begin__ = 27;\n"
       "        return stan::math::promote_scalar<fun_return_scalar_t__>(a6);\n"
       "    } catch (const std::exception& e) {\n"
-      "        stan::lang::rethrow_located(e, current_statement_begin__, "
+      "        stan::model::rethrow_located(e, current_statement_begin__, "
       "prog_reader__());\n"
       "        // Next line prevents compiler griping about no return\n"
       "        throw std::runtime_error(\"*** IF YOU SEE THIS, PLEASE REPORT A "
@@ -741,7 +741,7 @@ TEST(lang, local_function_args_hpp) {
       "        current_statement_begin__ = 31;\n"
       "        return stan::math::promote_scalar<fun_return_scalar_t__>(a7);\n"
       "    } catch (const std::exception& e) {\n"
-      "        stan::lang::rethrow_located(e, current_statement_begin__, "
+      "        stan::model::rethrow_located(e, current_statement_begin__, "
       "prog_reader__());\n"
       "        // Next line prevents compiler griping about no return\n"
       "        throw std::runtime_error(\"*** IF YOU SEE THIS, PLEASE REPORT A "
@@ -821,7 +821,7 @@ TEST(lang, local_function_args_hpp) {
       "        current_statement_begin__ = 35;\n"
       "        return stan::math::promote_scalar<fun_return_scalar_t__>(a8);\n"
       "    } catch (const std::exception& e) {\n"
-      "        stan::lang::rethrow_located(e, current_statement_begin__, "
+      "        stan::model::rethrow_located(e, current_statement_begin__, "
       "prog_reader__());\n"
       "        // Next line prevents compiler griping about no return\n"
       "        throw std::runtime_error(\"*** IF YOU SEE THIS, PLEASE REPORT A "
@@ -901,7 +901,7 @@ TEST(lang, local_function_args_hpp) {
       "        current_statement_begin__ = 39;\n"
       "        return stan::math::promote_scalar<fun_return_scalar_t__>(a9);\n"
       "    } catch (const std::exception& e) {\n"
-      "        stan::lang::rethrow_located(e, current_statement_begin__, "
+      "        stan::model::rethrow_located(e, current_statement_begin__, "
       "prog_reader__());\n"
       "        // Next line prevents compiler griping about no return\n"
       "        throw std::runtime_error(\"*** IF YOU SEE THIS, PLEASE REPORT A "
@@ -983,7 +983,7 @@ TEST(lang, local_function_args_hpp) {
       "        current_statement_begin__ = 43;\n"
       "        return stan::math::promote_scalar<fun_return_scalar_t__>(a10);\n"
       "    } catch (const std::exception& e) {\n"
-      "        stan::lang::rethrow_located(e, current_statement_begin__, "
+      "        stan::model::rethrow_located(e, current_statement_begin__, "
       "prog_reader__());\n"
       "        // Next line prevents compiler griping about no return\n"
       "        throw std::runtime_error(\"*** IF YOU SEE THIS, PLEASE REPORT A "
@@ -1067,7 +1067,7 @@ TEST(lang, local_function_args_hpp) {
       "        current_statement_begin__ = 47;\n"
       "        return stan::math::promote_scalar<fun_return_scalar_t__>(a11);\n"
       "    } catch (const std::exception& e) {\n"
-      "        stan::lang::rethrow_located(e, current_statement_begin__, "
+      "        stan::model::rethrow_located(e, current_statement_begin__, "
       "prog_reader__());\n"
       "        // Next line prevents compiler griping about no return\n"
       "        throw std::runtime_error(\"*** IF YOU SEE THIS, PLEASE REPORT A "
@@ -1151,7 +1151,7 @@ TEST(lang, local_function_args_hpp) {
       "        current_statement_begin__ = 51;\n"
       "        return stan::math::promote_scalar<fun_return_scalar_t__>(a12);\n"
       "    } catch (const std::exception& e) {\n"
-      "        stan::lang::rethrow_located(e, current_statement_begin__, "
+      "        stan::model::rethrow_located(e, current_statement_begin__, "
       "prog_reader__());\n"
       "        // Next line prevents compiler griping about no return\n"
       "        throw std::runtime_error(\"*** IF YOU SEE THIS, PLEASE REPORT A "
