+++ src/test/unit/lang/generator/multi_dim_var_test.cpp
@@ -168,7 +168,7 @@ TEST(lang, params_block_var_hpp_xform_inits) {
       "        current_statement_begin__ = 2;\n"
       "        if (!(context__.contains_r(\"ar_mat\")))\n"
       "            "
-      "stan::lang::rethrow_located(std::runtime_error(std::string(\"Variable "
+      "stan::model::rethrow_located(std::runtime_error(std::string(\"Variable "
       "ar_mat missing\")), current_statement_begin__, prog_reader__());\n"
       "        vals_r__ = context__.vals_r(\"ar_mat\");\n"
       "        pos__ = 0U;\n"
@@ -209,7 +209,7 @@ TEST(lang, params_block_var_hpp_xform_inits) {
       "ar_mat[i_0__][i_1__]);\n"
       "                } catch (const std::exception& e) {\n"
       "                    "
-      "stan::lang::rethrow_located(std::runtime_error(std::string(\"Error "
+      "stan::model::rethrow_located(std::runtime_error(std::string(\"Error "
       "transforming variable ar_mat: \") + e.what()), "
       "current_statement_begin__, prog_reader__());\n"
       "                }\n"
@@ -413,7 +413,7 @@ TEST(lang, transformed_params_block_var_hpp_log_prob) {
       "parameter: ar_mat\" << \"[\" << k_0__ << \"]\" << \"[\" << k_1__ << "
       "\"]\" << \"(\" << j_1__ << \", \" << j_2__ << \")\";\n"
       "                                "
-      "stan::lang::rethrow_located(std::runtime_error(std::string(\"Error "
+      "stan::model::rethrow_located(std::runtime_error(std::string(\"Error "
       "initializing variable ar_mat: \") + msg__.str()), "
       "current_statement_begin__, prog_reader__());\n"
       "                            }\n"
