+++ src/test/unit/lang/generator/prim_var_test.cpp
@@ -269,7 +269,7 @@ TEST(lang, params_block_var_hpp_xform_inits) {
       "        current_statement_begin__ = 2;\n"
       "        if (!(context__.contains_r(\"p2\")))\n"
       "            "
-      "stan::lang::rethrow_located(std::runtime_error(std::string(\"Variable "
+      "stan::model::rethrow_located(std::runtime_error(std::string(\"Variable "
       "p2 missing\")), current_statement_begin__, prog_reader__());\n"
       "        vals_r__ = context__.vals_r(\"p2\");\n"
       "        pos__ = 0U;\n"
@@ -281,7 +281,7 @@ TEST(lang, params_block_var_hpp_xform_inits) {
       "            writer__.scalar_unconstrain(p2);\n"
       "        } catch (const std::exception& e) {\n"
       "            "
-      "stan::lang::rethrow_located(std::runtime_error(std::string(\"Error "
+      "stan::model::rethrow_located(std::runtime_error(std::string(\"Error "
       "transforming variable p2: \") + e.what()), current_statement_begin__, "
       "prog_reader__());\n"
       "        }\n"
@@ -289,7 +289,7 @@ TEST(lang, params_block_var_hpp_xform_inits) {
       "        current_statement_begin__ = 3;\n"
       "        if (!(context__.contains_r(\"ar_p2\")))\n"
       "            "
-      "stan::lang::rethrow_located(std::runtime_error(std::string(\"Variable "
+      "stan::model::rethrow_located(std::runtime_error(std::string(\"Variable "
       "ar_p2 missing\")), current_statement_begin__, prog_reader__());\n"
       "        vals_r__ = context__.vals_r(\"ar_p2\");\n"
       "        pos__ = 0U;\n"
@@ -307,7 +307,7 @@ TEST(lang, params_block_var_hpp_xform_inits) {
       "                writer__.scalar_lub_unconstrain(0, 1, ar_p2[i_0__]);\n"
       "            } catch (const std::exception& e) {\n"
       "                "
-      "stan::lang::rethrow_located(std::runtime_error(std::string(\"Error "
+      "stan::model::rethrow_located(std::runtime_error(std::string(\"Error "
       "transforming variable ar_p2: \") + e.what()), "
       "current_statement_begin__, prog_reader__());\n"
       "            }\n"
@@ -316,7 +316,7 @@ TEST(lang, params_block_var_hpp_xform_inits) {
       "        current_statement_begin__ = 4;\n"
       "        if (!(context__.contains_r(\"ar_p3\")))\n"
       "            "
-      "stan::lang::rethrow_located(std::runtime_error(std::string(\"Variable "
+      "stan::model::rethrow_located(std::runtime_error(std::string(\"Variable "
       "ar_p3 missing\")), current_statement_begin__, prog_reader__());\n"
       "        vals_r__ = context__.vals_r(\"ar_p3\");\n"
       "        pos__ = 0U;\n"
@@ -335,7 +335,7 @@ TEST(lang, params_block_var_hpp_xform_inits) {
       "ar_p3[i_0__]);\n"
       "            } catch (const std::exception& e) {\n"
       "                "
-      "stan::lang::rethrow_located(std::runtime_error(std::string(\"Error "
+      "stan::model::rethrow_located(std::runtime_error(std::string(\"Error "
       "transforming variable ar_p3: \") + e.what()), "
       "current_statement_begin__, prog_reader__());\n"
       "            }\n"
@@ -512,7 +512,7 @@ TEST(lang, transformed_params_block_var_hpp_log_prob) {
       "                std::stringstream msg__;\n"
       "                msg__ << \"Undefined transformed parameter: p2\";\n"
       "                "
-      "stan::lang::rethrow_located(std::runtime_error(std::string(\"Error "
+      "stan::model::rethrow_located(std::runtime_error(std::string(\"Error "
       "initializing variable p2: \") + msg__.str()), "
       "current_statement_begin__, prog_reader__());\n"
       "            }\n"
@@ -524,7 +524,7 @@ TEST(lang, transformed_params_block_var_hpp_log_prob) {
       "                    msg__ << \"Undefined transformed parameter: ar_p2\" "
       "<< \"[\" << k_0__ << \"]\";\n"
       "                    "
-      "stan::lang::rethrow_located(std::runtime_error(std::string(\"Error "
+      "stan::model::rethrow_located(std::runtime_error(std::string(\"Error "
       "initializing variable ar_p2: \") + msg__.str()), "
       "current_statement_begin__, prog_reader__());\n"
       "                }\n"
